:recycle: Core usecases decoupled from api routes

+++ pkg/models/tag.go
@@ -19,6 +19,6 @@ type CreateTagParams struct {
 // Parameters to list all tags a user has used in the images
 type ListUserTagsParams struct {
 	ID     int64 `json:"id"`
-	Limit  int   `json:"limit"`
-	Offset int   `json:"offset"`
+	Limit  int64 `json:"limit"`
+	Offset int64 `json:"offset"`
 }
